Normalize selfhost bootstrap bundle surface
