+++ internal/selfhost/gen_selfhost.go
@@ -10,37 +10,9 @@ import (
 	"os/exec"
 	"path/filepath"
 	"strings"
+
+	"github.com/osty/osty/internal/selfhost/bundle"
 )
-
-var sourceFiles = []string{
-	"examples/selfhost-core/semver.osty",
-	"examples/selfhost-core/semver_parse.osty",
-	"examples/selfhost-core/frontend.osty",
-	"toolchain/lexer.osty",
-	"toolchain/parser.osty",
-	"examples/selfhost-core/formatter_ast.osty",
-	"examples/selfhost-core/check_bridge.osty",
-	"examples/selfhost-core/check.osty",
-	"examples/selfhost-core/resolve.osty",
-	"examples/selfhost-core/lint.osty",
-	"internal/selfhost/ast_lower.osty",
-}
-
-const stringsPrelude = `use go "strings" as strings {
-    fn Count(s: String, substr: String) -> Int
-    fn Fields(s: String) -> List<String>
-    fn HasPrefix(s: String, prefix: String) -> Bool
-    fn HasSuffix(s: String, suffix: String) -> Bool
-    fn Join(elems: List<String>, sep: String) -> String
-    fn ReplaceAll(s: String, old: String, new: String) -> String
-    fn Repeat(s: String, count: Int) -> String
-    fn Split(s: String, sep: String) -> List<String>
-    fn SplitN(s: String, sep: String, n: Int) -> List<String>
-    fn TrimPrefix(s: String, prefix: String) -> String
-    fn TrimSpace(s: String) -> String
-    fn TrimSuffix(s: String, suffix: String) -> String
-}
-`
 
 func main() {
 	if err := run(); err != nil {
@@ -66,7 +38,7 @@ func run() error {
 	}
 	defer os.RemoveAll(tmpDir)
 
-	merged, err := mergedSource(root)
+	merged, err := bundle.MergeSelfhost(root)
 	if err != nil {
 		return err
 	}
@@ -135,7 +107,7 @@ func generatedSelfhostUpToDate(root, outPath string) (bool, error) {
 	if err := checkPath(filepath.Join(root, "internal/selfhost/gen_selfhost.go")); err != nil {
 		return false, fmt.Errorf("stat selfhost generator: %w", err)
 	}
-	for _, rel := range sourceFiles {
+	for _, rel := range bundle.SelfhostSourceFiles {
 		if err := checkPath(filepath.Join(root, filepath.FromSlash(rel))); err != nil {
 			return false, fmt.Errorf("stat %s: %w", rel, err)
 		}
@@ -158,78 +130,6 @@ func findRepoRoot() (string, error) {
 		}
 		dir = parent
 	}
-}
-
-func mergedSource(root string) ([]byte, error) {
-	var b strings.Builder
-	b.WriteString(stringsPrelude)
-	b.WriteByte('\n')
-	for _, rel := range sourceFiles {
-		path := filepath.Join(root, filepath.FromSlash(rel))
-		data, err := os.ReadFile(path)
-		if err != nil {
-			return nil, fmt.Errorf("read %s: %w", rel, err)
-		}
-		b.WriteString("// ---- ")
-		b.WriteString(rel)
-		b.WriteString(" ----\n")
-		trimmed := stripLeadingStringsUse(string(data))
-		trimmed = normalizeStdStringsCalls(trimmed)
-		b.WriteString(trimmed)
-		if !strings.HasSuffix(trimmed, "\n") {
-			b.WriteByte('\n')
-		}
-		b.WriteByte('\n')
-	}
-	return []byte(b.String()), nil
-}
-
-func stripLeadingStringsUse(src string) string {
-	const goPrefix = `use go "strings" as strings {`
-	const stdPrefix = "use std.strings as strings"
-
-	lines := strings.SplitAfter(src, "\n")
-	for i := 0; i < len(lines); i++ {
-		trimmed := strings.TrimSpace(lines[i])
-		switch {
-		case strings.HasPrefix(trimmed, goPrefix):
-			for j := i + 1; j < len(lines); j++ {
-				if strings.TrimSpace(lines[j]) == "}" {
-					return strings.Join(lines[:i], "") + strings.Join(lines[j+1:], "")
-				}
-			}
-			return src
-		case trimmed == stdPrefix:
-			return strings.Join(lines[:i], "") + strings.Join(lines[i+1:], "")
-		}
-	}
-
-	return src
-}
-
-// normalizeStdStringsCalls rewrites camelCase strings.xxx() calls (from
-// std.strings) to PascalCase so they match the Go FFI declarations in
-// stringsPrelude. Only applies to files that originally used std.strings.
-func normalizeStdStringsCalls(src string) string {
-	// Only touch files that had the std.strings import (now stripped).
-	// We detect this by checking whether any camelCase strings.* call appears.
-	for _, pair := range [][2]string{
-		{"strings.split(", "strings.Split("},
-		{"strings.join(", "strings.Join("},
-		{"strings.hasPrefix(", "strings.HasPrefix("},
-		{"strings.hasSuffix(", "strings.HasSuffix("},
-		{"strings.trimSpace(", "strings.TrimSpace("},
-		{"strings.replaceAll(", "strings.ReplaceAll("},
-		{"strings.trimPrefix(", "strings.TrimPrefix("},
-		{"strings.trimSuffix(", "strings.TrimSuffix("},
-		{"strings.repeat(", "strings.Repeat("},
-		{"strings.count(", "strings.Count("},
-		{"strings.fields(", "strings.Fields("},
-		{"strings.splitN(", "strings.SplitN("},
-	} {
-		src = strings.ReplaceAll(src, pair[0], pair[1])
-	}
-	return src
 }
 
 func patchGenerated(path string) error {
